requirements.txt
beautifulsoup4==4.11.1
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
colorama==0.4.5
filelock==3.7.1
greenlet==1.1.2
html5lib==1.1
huggingface-hub==0.8.1
idna==3.3
joblib==1.1.0
nltk==3.7
numpy==1.23.1
packaging==21.3
pandas==1.4.3
psycopg2==2.9.3
pyparsing==3.0.9
python-dateutil==2.8.2
python-decouple==3.6
pytz==2022.1
PyYAML==6.0
regex==2022.7.9
requests==2.28.1
six==1.16.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.39
tokenizers==0.12.1
tqdm==4.64.0
transformers==4.20.1
typing_extensions==4.3.0
urllib3==1.26.10
webencodings==0.5.1
